+++ AllMyLights/Platforms/AutostartEntry.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 using AllMyLights.Platforms.Linux;
+using AllMyLights.Platforms.Windows;
 using NLog;
 
 namespace AllMyLights.Platforms
@@ -12,6 +13,7 @@ namespace AllMyLights.Platforms
         public static AutostartEntry GetPlatformInstance() 
         {
             if (OperatingSystem.IsLinux()) return new SystemdAutostartEntry();
+            if (OperatingSystem.IsWindows()) return new WindowsAutostartEntry();
 
             Logger.Error($"{nameof(AutostartEntry)} not implemented for {RuntimeInformation.OSDescription}");
             Environment.Exit((int)ExitCode.PlatformNotSupported);